refactor: Change name of project card
